+++ cavgym/scenarios/bus_stop.py
@@ -7,10 +7,10 @@ from cavgym.scenarios import car_constants, bicycle_constants, bus_constants
 road_map = RoadMap(
     major_road=Road(
         constants=RoadConstants(
-            length=1600.0,
+            length=1600,
             num_outbound_lanes=3,
             num_inbound_lanes=0,
-            lane_width=40.0,
+            lane_width=40,
             position=utilities.Point(0.0, 0.0),
             orientation=utilities.DEG2RAD * 0.0
         )
@@ -18,8 +18,8 @@ road_map = RoadMap(
 )
 
 env_constants = CAVEnvConstants(
-    viewer_width=int(road_map.major_road.constants.length),
-    viewer_height=int(road_map.major_road.constants.lane_width * (road_map.major_road.constants.num_outbound_lanes + road_map.major_road.constants.num_inbound_lanes + 2)),
+    viewer_width=road_map.major_road.constants.length,
+    viewer_height=road_map.major_road.constants.lane_width * (road_map.major_road.constants.num_outbound_lanes + road_map.major_road.constants.num_inbound_lanes + 2),
     time_resolution=1.0 / 60.0,
     road_map=road_map
 )
@@ -30,7 +30,7 @@ actors = [
             position=road_map.major_road.outbound.lane_spawns[0],
             velocity=100.0,
             orientation=road_map.major_road.outbound_orientation,
-            acceleration=0.0,
+            acceleration=0,
             angular_velocity=0.0
         ),
         constants=bus_constants
@@ -40,7 +40,7 @@ actors = [
             position=road_map.major_road.outbound.lane_spawns[1],
             velocity=100.0,
             orientation=road_map.major_road.outbound_orientation,
-            acceleration=0.0,
+            acceleration=0,
             angular_velocity=0.0
         ),
         constants=car_constants
@@ -50,7 +50,7 @@ actors = [
             position=road_map.major_road.outbound.lane_spawns[2],
             velocity=100.0,
             orientation=road_map.major_road.outbound_orientation,
-            acceleration=0.0,
+            acceleration=0,
             angular_velocity=0.0
         ),
         constants=bicycle_constants
